Restructuring plugins for Neovim by placing each plugin into its respective category subdirectory within a parent plugins/ directory. Moving configuration Lua files to a new config/ directory
