+++ nvim/lua/config/keybindings.lua
