+++ nvim/lua/config/lazy.lua
@@ -1,0 +1,26 @@
+-- This module calls lazy.nvim to load all plugins specified in the plugins/ directory.
+
+-- Bootstrap lazy.nvim
+local lazypath = vim.fn.stdpath("data") .. "/lazy/lazy.nvim"
+if not (vim.uv or vim.loop).fs_stat(lazypath) then
+	local lazyrepo = "https://github.com/folke/lazy.nvim.git"
+	local out = vim.fn.system({ "git", "clone", "--filter=blob:none", "--branch=stable", lazyrepo, lazypath })
+
+	if vim.v.shell_error ~= 0 then
+		vim.api.nvim_echo({
+			{ "Failed to clone lazy.nvim:\n", "ErrorMsg" },
+			{ out,                            "WarningMsg" },
+			{ "\nPress any key to exit..." },
+		}, true, {})
+
+		vim.fn.getchar()
+
+		os.exit(1)
+	end
+end
+
+vim.opt.rtp:prepend(lazypath)
+
+print("CALLINNG LAZY SETUP PLUGINS")
+
+require("lazy").setup("plugins")
