+++ nvim/lua/plugins/coding/copilot.lua
@@ -1,0 +1,1 @@
+return { "github/copilot.vim" }
