+++ nvim/lua/plugins/coding/nvim-dap.lua
@@ -1,0 +1,1 @@
+return { "mfussenegger/nvim-dap" }
