+++ nvim/lua/plugins/coding/nvim-lint.lua
@@ -1,0 +1,1 @@
+return { "mfussenegger/nvim-lint" }
