+++ nvim/lua/plugins/coding/vim-gitgutter.lua
@@ -1,0 +1,1 @@
+return { "airblade/vim-gitgutter" }
