+++ nvim/lua/plugins/formatting/indent-blankline.lua
@@ -1,0 +1,11 @@
+return {
+	"lukas-reineke/indent-blankline.nvim",
+	config = function()
+		require("ibl").setup()
+	end,
+	lazy = false,
+	main = "ibl",
+	opts = {
+		char = "┊",
+	}
+}
