+++ nvim/lua/plugins/formatting/nerdcommenter.lua
@@ -1,0 +1,1 @@
+return { "preservim/nerdcommenter" }
