+++ nvim/lua/plugins/productivity/obsidian.lua
@@ -1,0 +1,7 @@
+return {
+	"epwalsh/obsidian.nvim",
+	event = { "VeryLazy" },
+	ft = "markdown",
+	--lazy = true,
+	version = "*"
+}
