+++ nvim/lua/plugins/themes/gruvbox.lua
@@ -1,0 +1,1 @@
+return { "ellisonleao/gruvbox.nvim" }
