+++ nvim/lua/plugins/themes/kanagawa.lua
@@ -1,0 +1,39 @@
+return {
+	"rebelot/kanagawa.nvim",
+	config = function()
+		require("kanagawa").setup({
+			colors = {
+				theme = {
+					all = {
+						ui = {
+							bg_gutter = "none"
+						}
+					}
+				}
+			},
+			overrides = function(colors)
+				local theme = colors.theme
+				return {
+					-- Set more uniform colors for the popup menu.
+					Pmenu = { fg = theme.ui.shade0, bg = theme.ui.bg_p1 },
+					PmenuSel = { fg = "NONE", bg = theme.ui.bg_p2 },
+					PmenuSbar = { bg = theme.ui.bg_m1 },
+					PmenuThumb = { bg = theme.ui.bg_p2 },
+
+					-- Override configurations for `Telescope`.
+					TelescopeTitle = { fg = theme.ui.special, bold = true },
+					TelescopePromptNormal = { bg = theme.ui.bg_p1 },
+					TelescopePromptBorder = { fg = theme.ui.bg_p1, bg = theme.ui.bg_p1 },
+					TelescopeResultsNormal = { fg = theme.ui.fg_dim, bg = theme.ui.bg_m1 },
+					TelescopeResultsBorder = { fg = theme.ui.bg_m1, bg = theme.ui.bg_m1 },
+					TelescopePreviewNormal = { bg = theme.ui.bg_dim },
+					TelescopePreviewBorder = { bg = theme.ui.bg_dim, fg = theme.ui.bg_dim },
+				}
+			end,
+		})
+
+		vim.cmd("colorscheme kanagawa")
+	end,
+	lazy = false,
+	priority = 10000
+}
