+++ nvim/lua/plugins/themes/nord.lua
@@ -1,0 +1,6 @@
+return {
+	"shaunsingh/nord.nvim",
+	--config = function()
+	--vim.cmd("colorscheme nord")
+	--end
+}
