+++ nvim/lua/plugins/ui/zen-mode.lua
@@ -1,0 +1,9 @@
+return {
+	"folke/zen-mode.nvim",
+	lazy = false,
+	opts = {
+		window = {
+			width = 140
+		}
+	}
+}
